requirements.txt

pdfplumber
customtkinter
sv_ttk
darkdetect
Pillow